Added an option to force monospace text rendering
Uses the <tt> HTML tag, and if turned on, forces the tag's use on all rendered messages. It's kind of blunt force, but it renders ASCII art better than the normal renderer, and some users may want it.

New setting is added to the config file, and the "messages" section of the settings dialog.