Changed windows menu
Windows menu now looks more like the systray windows menu, and adds an entry to open the server channel search.

Appearance of the channel search option in the windows menu can be toggled on an off in the config file/module and the settings dialog. It is turned on by default