Added functions for future functionality
I found a bug; channels with "messed up" names (think #/ or #:coffee:, and yes, that's an emoji) cannot be written to log at the current moment. I'm thinking about using url-safe base65 to en/decode log filenames so that we can still log these channels. The functions added to merk/resources.py are for that.