Changes to log and style filenames
Log and style filenames are now escaped for "special" characters (i.e., ones that can't be used in a filename). Filenames are backwards compatible with older versions of MERK.

This fixes a bug with channels that used "special"characters in their name (specifically, I discovered this when I joined the #/ channel on EFnet). Logs and styles for these channels can now be saved, loaded, and exported.